Enable index only scan on ao/aocs table
The index scan has disabled on ao/aocs table for that
random access performance on the ao/aocs table is extremely
low. But index only scan does not access data files. Its
performance should be as good as the heap table. So index
only scan should be enable on ao/aocs table.

The san node use eof in aoseg table and visibility map to
exam the visibility of tuples in index. The member variables
lastRowNum and firstRowNum are introduced into AOCSFetchDesc
/ AppendOnlyFetchDesc to describe the range of visible row
numbers. Then use the visibility map to check the visibility
of each tuple.