Use asynchronous connection in libpqwalreceiver
The previous version, which used poll()/select(), caused some problems. For example,
if the master is no longer available, we call gpactivatestandby while the walreceiver
is trying to connect to the master. The walreceiver receives SIGTERM, but the poll()
continues its work. gpactivatestandby tries to connect to new master, but it fails due
to an expired timeout. To avoid such errors, WaitLatchOrSocket() is used, which is
more portable and flexible. The implementation logic is taken from newer versions of
Postgres (refer to commit https://github.com/postgres/postgres/commit/1e8a850094 and
https://github.com/postgres/postgres/commit/597a87ccc9).

Also change walreceiver to use its procLatch instead of a custom latch.