Use asynchronous connection in libpqwalreceiver
The previous version, which used poll()/select(), caused some problems. For example,
if the master is no longer available, we call gpactivatestandby while the walreceiver
is trying to connect to the master. The walreceiver receives SIGTERM, but the poll()
continues its work. gpactivatestandby tries to connect to new master, but it fails due
to an expired timeout. To avoid such errors, WaitLatchOrSocket() is used, which is
more portable and flexible. The implementation is aligned with Upstream REL_12_BETA2,
and the logic is taken from newer versions of Postgres commit b5dd50f and a1a789e.