Prevent REINDEX TABLE partitioned table statement in function.
As different from upstream, we support REINDEX TABLE on partitioned
table which will expand the partitioned table and start new transaction
to do the reindex for each table. This has side effects that cannot be
rolled back if it gets called in PL language and may crash, so we need
to make sure it must not run inside a function.
This fix is related to issue https://github.com/greenplum-db/gpdb/issues/11948.

This is applied to STABLE release only for less impact of current transaction
behavior. It will be changed in the next major release to fully address
preventing multi-transaction statement running in a transaction block.

Co-authored-by: Haolin Wang <whaolin@vmware.com>