On segments balance syncrep off and wal catchup when mirror wal is catching up.
In real scenarios it often happens that the mirror node is down (probably not
really down) and out-of-sync for some time and that could lead to large wal lag
on the mirror node, and when the mirror is added back by fts, fts currently
soon sets syncrep on. While at the time the mirror is catching up and thus it
stalls the new write query for a long time until wal catchup finishes.

There was a guc repl_catchup_within_range that "Sets the maximum number of xlog
segments allowed to lag when the backends can start blocking despite the WAL
sender being in catchup phase." It was for the coordinator only. Now we could
use it on segments also, but this mechanism is not sufficient if the new wal is
generated faster - wal catchup would need more time and the primary is at the
risk of data loss during the period due to syncrep off. So we do further
throttle in large wal write function to slow down those transactions a bit and
leave more resources to wal catchup.

xlog lag is calculated using GetFlushRecPtr() -
XLogGetReplicationSlotMinimumLSN(), but note XLogGetReplicationSlotMinimumLSN()
is not updated very frequently - I created a thead about this in pgsql-hackers
mail list but this does not affect our patch much. (just slows down the write
query further during wal catchup).

A simple bulk insert testing shows during catchup it reduces the catchup time
by 50%, while the query running time increases by 60%. (Note in my testing, I
added 10ms after each walrcv_receive() since I was testing in a demo cluster
that uses ver fast loopback connection so catchup was very fast).

Need to consider the scenario when catchup is suspended - that should
not stall the write operationn.

Co-authored-by: Haolin Wang <whaolin@vmware.com>