Add throttle logic to slow down WAL writes in CATCHUP phase.
There was a guc repl_catchup_within_range that "Sets the maximum number of xlog
segments allowed to lag when the backends can start blocking despite the WAL
sender being in catchup phase." It was for the coordinator only. Now we could
use it on segments also, but this mechanism is not sufficient if the new wal is
generated faster - wal catchup would need more time and the primary is at the
risk of data loss during the period due to syncrep off. So we do further
throttle in large wal write function to slow down those transactions a bit and
leave more resources to wal catchup.

xlog lag is calculated using GetFlushRecPtr() -
XLogGetReplicationSlotMinimumLSN(), but note XLogGetReplicationSlotMinimumLSN()
is not updated very frequently - I created a thead about this in pgsql-hackers
mail list but this does not affect our patch much. (just slows down the write
query further during wal catchup).

Co-authored-by: Haolin Wang <whaolin@vmware.com>