Commits start blocking only if STREAMING or CATCHUP within range.
In real scenarios it often happens that the mirror node is down (probably not
really down) and out-of-sync for some time and that could lead to large wal lag
on the mirror node, and when the mirror is added back by fts, fts currently
soon sets syncrep on. While at the time the mirror is catching up and thus it
stalls the new write query for a long time until wal catchup finishes.

There was a guc repl_catchup_within_range that "Sets the maximum number of xlog
segments allowed to lag when the backends can start blocking despite the WAL
sender being in catchup phase." It was for the coordinator only. Now we could
use it on segments also, to allow commits start blocking only if walsender is
in STREAMING state or in CATCHUP state but repl_catchup_within_range.

Co-authored-by: Haolin Wang <whaolin@vmware.com>