[6X] Fix AO fsync PANIC when creating restart point during crash recovery by gprecoverseg on old primary segments.
The patch is intended to fix AO table fsync PANIC during crash recovery.

The scenario is when a checkpoint was executing on the primary segment,
a bunch of insert requests were received so generate a lot of insertion
XLOG records between the REO point and the LSN of the current checkpoint.
The primary segment then crashed after the checkpoint completed.
After the mirror segment was promoted to new primary, a truncate request
was received to truncate this AO table. Then both old relfilenode and its
segment files were removed when executing gprecoverseg.
On the old primary (new mirror), those files were also removed but only
the old refilenode segment files would be re-generated due to the insert
records during WAL replay. This would queue fsync request to checkpointer
and would not be forgotten in later flush. Later restart point creation would
flush the buffer to disk and expect to access the AO old base relfilenode
file (.0), while it was removed by pg_rewind previously, therefore PANIC
occured at mdsync during creating restart point.

The fix is to remove the logic of accessing the base relfilenode file
when the target is an AO segment file in the above fsync scenario.
Also, instead of reporting the segment file name, it will correct to report
the right file name for ENOENT if the base relfilenode file can't be found.

The issue wasn't found on master branch since the AO fsync interface is
different from Heap table and it works as expected as the above fix.