Vacuum index on Append-Optimized table enhancement.
In the past, we rely on the visibility map to check if an index entry
can be removed. Scanning VM is not such effective. With this new strategy,
we do index vacuum iff after we recycle a PendingDrop segment, and
think the index entry is removable only check if it points to a dropable
segment, so the checking would be effective, and dropable segments also
guarantee the index entry is invisible to anyone already.