Resolve GPDB_12_MERGE_FIXME: Not scan visimap for AO/AOCS costing
For Heap table, visibility map may help to estimate
the number of page fetches can be avoided during an
index-only scan. That is not the case for AO/AOCS table
since index-only scan hasn't been used with AO/AOCS.
So leave the comment here for future reference once
we have a clear requirement to do that.