Tolerate failure of bad buffer flushing on a temporary table.
This is intended to fix the issue #13809 (see also #6351).
The problem is, when the writer aborts the transaction before
any of readers, it drops the shared_buffer and unlinks corresponding
temporary relfilenodes. But readers may be yet to receive the cancel
signal from QD hence continue executing their part of the plan. This
makes a reader has a chance to read previous unlinked relfilenode
into shared_buffers and re-mark it to dirty in the case of hintbit is
set, which could result to a permanent failure could not open file
when other processes (such like bgworker or readers) attempt to evict
or flush this dirty buffer to disk.

The patch is trying to address a kinda like semantic contradiction
that a temporary object should not generate a permanent failure to
block later queries permanently. Given above context, the patch clears
the dirty shared buffer of the non-existent temporary relfilenode to
allow queries to be proceeded as is. Those queries could be finally
failed because of accessing that non-existent file but that just
generates a transient failure rather than a permanent failure, hence
DB instance could be continue working.