brin ao/co: Fix desummarization (utility mode)
This commit fixes desummarization for AO/CO tables by positioning the AO
revmap iterator correctly on to the block sequence chain, in which the
supplied heap block falls.

Caution: brin_desummarize_range() is not yet MPPized - it doesn't
dispatch yet (and its failure is masked by its void return type). So, we
test the functionality in utility mode. MPPizing it will be deferred to
a future commit.