ao/co: Implement partial scans
BRIN summarization invokes the table AM API index_build_range_scan() to
partially scan block ranges. For AO/CO tables, unfortunately, this
feature wasn't fully implemented - we used to scan the entire AO/CO
table, with each partial scan request, and discard the rows that didn't
belong to the desired range. This was a FIXME in the code and caused
untenable performance such as in #14836.

Now, we implement partial scans of AO/CO tables, by taking advantage of
the block directory to look up a safe (segment, offset) from which we
can kick off the scan. To determine where the scan ends, we look at the
logical heap block number of the tuple returned, and see if it lies in
the desired range. In case it falls in a hole, we conservatively start
scanning from the blkdir entry prior to the hole (and discard unneeded
rows).