Add AOTupleTableSlot and TTSOpsAOTuple
We used to immediately call `memtuple_deform` in
`AppendOnlyExecutorReadBlock_ProcessTuple` after
scanning an AO tuple to convert the `memtuple` into
`slot->data`. To JIT accelerate AO tuple deform,
this commit introduces `AOTupleTableSlot` and
`TTSOpsAOTuple` to store `memtuple` and
`MemTupleBinding`, which are then deformed in
`ExecInterpExpr`. This is consistent with heap
tuple deform.