Support missing-mode ADD COLUMN for AOCO tables
This commit optimizes AT ADD COLUMN for ao_column tables in the similar way as
ao_row tables: the new column will be regarded as "missing" for all existing
rows at the time of ADD COLUMN. We record the "last row numbers" of each segment
file for the new column. At table scanning time, we will use that information to
decide if a certain row is missing. If so, we'll fetch the default value from
pg_attribute.attmissingval instead of the file.

However, different than ao_row tables, for ao_column tables we cannot completely
avoid writing: columns files in ao_column tables are scanned individually. So
they have to have information about what rows exist in them. Therefore, during
ADD COLUMN, we still need to scan an existing column, and write same number of
rows in the new column. Currently, we are writing a NULL in place of the missing
value. Therefore, this optimization is more useful when the column's default
value is large. E.g.:

```
-- prep
create table co(a int) using ao_column;
insert into co select * from generate_series(1,10000000);

-- w/o the optimization
postgres=# alter table co add column b text default repeat('abc',100);
ALTER TABLE
Time: 11229.041 ms (00:11.229)

-- w/ the optimization
postgres=# alter table co add column b text default repeat('abc',100);
ALTER TABLE
Time: 477.533 ms
```

We do not simply write nothing because it would be a far more invasive change
as there are too many expectations in the code for the block header to be
consistent with the block content. It would be hard to get it right. So we would
want to wait and see if there's a real need to invest the time on it.

Also fixed an issue that we should allow alter column constraint to be done
with alter column optimization.

Co-authored-by: Divyesh Vanjare <vanjared@vmware.com>
Co-authored-by: Huansong Fu <fuhuansong@gmail.com>