Implement AO visimap cache
This commit implements an AO visimap cache for improving speed
of query on pg_aovisimap. It uses {segno, firstrowno} as the key
of cache entries and is hash-mapped to a rangeid which is managed
by a simple LRU array. The rangeid identifies a range of consecutive
rowNums (max 32768 rows), is used as the index of the LRU array.
If all rows in a range are visible, allvisible will be set to true
for this range or false otherwise.

The AO visimap cache lookup in visibility check context is:
    - checks cache.lastrange to see if maching current range identified by
      the {segno, firstrowno} (calculated based on the given AOTupleId),
      if matched and the range.allvisible is true, returns true immediately;
    - otherwise, searches the corresponding rangeid from the cache.hashtable,
      then get the LRU element indexed by the rangeid to check the allvisible
      flag of the range.

The max number of cache entries is controlled by GUC gp_aovisimap_max_cache_entries,
once the number is reached, LRU array evicts the least recently used element for
making room for new rangeid (its corresponding hash entry will be removed as well).