Write nextGxid as-is in checkpoint
Currently, when creating the checkpoint, we bump the nextGxid by the value of
GxidCount and write nextGxid in the checkpoint record. This is for the purpose
of skipping possibly-recovering dtx during restart. It works fine until we want
to support hot standby. In a hot standby, the latestCompletedGxid needs to be
properly initialized in order for the distributed snapshot to work. Similar to
the relationship between latestCompletedXid and nextFullXid, latestCompletedGxid
is initialized to nextGxid. However, since the standby sees a much advanced
value of nextGxid (after the gxid bump), in the case where standby restarted but
the primary has not, the primary might be still using some gxid that the
standby thinks are completed (since latestCompletedGxid has moved past them),
which could lead to wrong results.

Now, we do not bump nextGxid in checkpoint, but write the value as-is. To still
cover dtx recvoery, we bump it during the restart of the primary. We bump it
by the maximum value of gxid bump to make sure we've jumped over all of them.

We don't need to change BootstrapXLOG() as dtx recovery isn't an issue there.