Fix wrong values returned on sampling CO tuples.
This is intended to fix the issue that ANALYZE AO_COLUMN
table returned dead tupe values for non-first attributes.

An example case was mentioned in commit message 9f1e865.

The RCA of this issue is the internal variable
`blockRowsProcessed` of non-first attributes
missed update due to skipping processing on dead
tuples, leading to a lag of row offset comparing
to 1st column.

The solution is align the 1st column with others
that not update `blockRowsProcessed` on dead tuples
until next time advance, and keep skipping dead
tuples processing.